test: extract dummy icmp6 utils for tests
This extracts common implementation of dummy icmp6 utils used by tests.
