+++ src/libsystemd-network/meson.build
@@ -82,10 +82,16 @@ executables += [
                 'sources' : files('test-lldp-rx.c'),
         },
         network_test_template + {
-                'sources' : files('test-ndisc-ra.c'),
+                'sources' : files(
+                        'test-ndisc-ra.c',
+                        'icmp6-util-unix.c',
+                ),
         },
         network_test_template + {
-                'sources' : files('test-ndisc-rs.c'),
+                'sources' : files(
+                        'test-ndisc-rs.c',
+                        'icmp6-util-unix.c',
+                ),
         },
         network_test_template + {
                 'sources' : files('test-sd-dhcp-lease.c'),
@@ -106,6 +112,9 @@ executables += [
                 'sources' : files('fuzz-lldp-rx.c'),
         },
         network_fuzz_template + {
-                'sources' : files('fuzz-ndisc-rs.c'),
+                'sources' : files(
+                        'fuzz-ndisc-rs.c',
+                        'icmp6-util-unix.c',
+                ),
         },
 ]
